- Added engineer controls to manage the list view for the users tab. - Added the Engineer class - Created a new database for the information related to engineers. - Created the default user view that shows a list of all the engineers in the database. - Added and styled a new user window. Implemented the functionality that allows new users to be saved into the database alongside their details.
Summary: You can now click on the users tab in the main window in order to see a list of all of the engineers stored in the database. By selecting one of the engineers in the list and clicking the update button, you will then be welcomed with a window that allows you to update the details of the engineer. Clicking save will store this information in the database which will reflect in the user listview.